Move Groups feature to vertical-slice folders
Pilot continues with the Groups slice (2 endpoints — list videos
across all groups the user belongs to, list videos for a single
group). Controller, service, interface, DTOs, route constants, and
tests now live under Features/Groups/ in their respective projects.

- IGroupService moved out of Domain.Services into
  Application.Features.Groups.
- GroupRoutes extracted from ApiEndpoints.cs.
- DI registration moved into per-feature AddGroupsFeature()
  extension in GroupsModule.cs.
- Contracts split: GroupWithVideosResponse moves into
  Contracts/Features/Groups/, while VideoInformationResponse and
  VideoInformationModel (used across the Videos/Events slices)
  stay in Contracts/Responses/ in a new dedicated file until those
  slices migrate.
- EventsController gains a using for Application.Features.Groups
  since it still consumes IGroupService.
- Mobile (Library, MAUI Video model, test project) gain a using
  for the new Contracts.Features.Groups namespace alongside their
  existing Contracts.Requests/Responses usings.
