+++ src/backend/Application/Features/Groups/GroupService.cs
@@ -1,8 +1,7 @@
 ﻿using Domain.Entities;
-using Domain.Services;
 using Microsoft.EntityFrameworkCore;
 
-namespace Application.Services;
+namespace Application.Features.Groups;
 
 public class GroupService : IGroupService
 {
