+++ src/backend/Application/Features/Groups/GroupsModule.cs
@@ -1,0 +1,15 @@
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Application.Features.Groups;
+
+public static class GroupsModule
+{
+    extension(IServiceCollection services)
+    {
+        public IServiceCollection AddGroupsFeature()
+        {
+            services.AddScoped<IGroupService, GroupService>();
+            return services;
+        }
+    }
+}
