+++ src/backend/TB.DanceDance.API.Contracts/Features/Groups/GroupWithVideosResponse.cs
@@ -1,0 +1,15 @@
+using System;
+using System.Collections.Generic;
+using TB.DanceDance.API.Contracts.Responses;
+
+namespace TB.DanceDance.API.Contracts.Features.Groups
+{
+    public class GroupWithVideosResponse
+    {
+        public Guid GroupId { get; set; }
+        public string GroupName { get; set; } = null!;
+        public DateTime SeasonStart { get; set; }
+        public DateTime SeasonEnd { get; set; }
+        public ICollection<VideoInformationModel> Videos { get; set; } = null!;
+    }
+}
