+++ src/backend/TB.DanceDance.API/Features/Groups/GroupRoutes.cs
@@ -1,0 +1,10 @@
+namespace TB.DanceDance.API.Features.Groups;
+
+public static class GroupRoutes
+{
+    private const string ApiBase = "api";
+    private const string Base = $"{ApiBase}/groups";
+
+    public const string Videos = $"{Base}/videos";
+    public const string VideosForGroup = $"{Base}/{{groupId:guid}}/videos";
+}
