+++ src/mobile/TB.DanceDance.Mobile.Library/Data/Models/Video.cs
@@ -1,4 +1,5 @@
-﻿using TB.DanceDance.API.Contracts.Responses;
+﻿using TB.DanceDance.API.Contracts.Features.Groups;
+using TB.DanceDance.API.Contracts.Responses;
 
 namespace TB.DanceDance.Mobile.Library.Data.Models;
