+++ src/mobile/TB.DanceDance.Mobile.Library/Services/DanceApi/IDanceHttpApiClient.cs
@@ -1,4 +1,5 @@
-﻿using TB.DanceDance.API.Contracts.Models;
+﻿using TB.DanceDance.API.Contracts.Features.Groups;
+using TB.DanceDance.API.Contracts.Models;
 using TB.DanceDance.API.Contracts.Requests;
 using TB.DanceDance.API.Contracts.Responses;
