+++ src/tests/TB.DanceDance.Mobile.Tests/IntegrationTests/DanceHttpApiClientTests.cs
@@ -1,5 +1,6 @@
 ﻿using NSubstitute;
 using System.Text.Json;
+using TB.DanceDance.API.Contracts.Features.Groups;
 using TB.DanceDance.API.Contracts.Models;
 using TB.DanceDance.API.Contracts.Requests;
 using TB.DanceDance.API.Contracts.Responses;
