+++ src/tests/TB.DanceDance.Tests/Features/Groups/GroupServiceTests.cs
@@ -1,9 +1,9 @@
-﻿using Application.Services;
+﻿using Application.Features.Groups;
 using Domain.Entities;
 using Infrastructure.Data;
 using TB.DanceDance.Tests.TestsFixture;
 
-namespace TB.DanceDance.Tests.Application;
+namespace TB.DanceDance.Tests.Features.Groups;
 
 public class GroupServiceTests : BaseTestClass
 {
